requirements.txt
langgraph
langchain
python-dotenv
langchain-ollama
Ipython
gradio
langchain-openai
langchain-ollama
redis
Ipython
langchain-redis
